Implemented forgiving jump detection.
This is in the form of a short delay after leaving the ground in which the player can still jump. The number of frames that the player can still jump is measured by m_JumpBufferFrameAmount. This helps greatly when the player hits small bumps and leaves the ground, and finds their jump inputs 'disappearing' on them.