Fixed Ascendant Snagging and did some animation work.
Ascendant Snagging is when the player's foot hits a surface that is less vertical than its current trajectory, and gets pulled downward onto the flatter slope. The code detects this circumstance by comparing the new surface's y value to the y value of the player trajectory. If the player's y value is higher, they are hitting the surface from the underside with their foot collider, which should be ignored.